refactor(core): Use kork-expressions (#3071)
* Use kork-expressions

* Refactor ExpressionFunctionProvider to reduce collections boilerplate

* Support yaml parsing
