+++ orca-core/src/test/groovy/com/netflix/spinnaker/orca/pipeline/expressions/functions/StageExpressionFunctionProviderSpec.groovy
@@ -16,14 +16,13 @@
 
 package com.netflix.spinnaker.orca.pipeline.expressions.functions
 
+import com.netflix.spinnaker.kork.expressions.SpelHelperFunctionException
 import com.netflix.spinnaker.orca.ExecutionContext
-import com.netflix.spinnaker.orca.pipeline.expressions.SpelHelperFunctionException
 import spock.lang.Shared
 import spock.lang.Specification
 import spock.lang.Unroll
 
-import static com.netflix.spinnaker.orca.pipeline.expressions.functions.StageExpressionFunctionProvider.currentStage
-import static com.netflix.spinnaker.orca.pipeline.expressions.functions.StageExpressionFunctionProvider.stageByRefId
+import static com.netflix.spinnaker.orca.ExecutionStatus.SUCCEEDED
 import static com.netflix.spinnaker.orca.test.model.ExecutionBuilder.pipeline
 import static com.netflix.spinnaker.orca.test.model.ExecutionBuilder.stage
 
@@ -34,12 +33,76 @@ class StageExpressionFunctionProviderSpec extends Specification {
       id = "1"
       refId = "1.0"
       name = "My First Stage"
+      context = [
+        "region": "us-east-1",
+      ]
     }
 
     stage {
       id = "2"
       refId = "2.0"
       name = "My Second Stage"
+      context = [
+        "region": "us-west-1",
+      ]
+    }
+
+    stage {
+      id = "3"
+      refId = "3"
+      status = SUCCEEDED
+      type = "createServerGroup"
+      name = "Deploy in us-east-1"
+      context.putAll(
+        "account": "test",
+        "deploy.account.name": "test",
+        "availabilityZones": [
+          "us-east-1": [
+            "us-east-1c",
+            "us-east-1d",
+            "us-east-1e"
+          ]
+        ],
+        "capacity": [
+          "desired": 1,
+          "max"    : 1,
+          "min"    : 1
+        ],
+        "deploy.server.groups": [
+          "us-east-1": [
+            "app-test-v001"
+          ]
+        ]
+      )
+    }
+
+    stage {
+      id = "4"
+      refId = "4"
+      status = SUCCEEDED
+      type = "disableServerGroup"
+      name = "disable server group"
+      context.putAll(
+        "account": "test",
+        "deploy.account.name": "test",
+        "availabilityZones": [
+          "us-east-1": [
+            "us-east-1c",
+            "us-east-1d",
+            "us-east-1e"
+          ]
+        ],
+        "capacity": [
+          "desired": 1,
+          "max"    : 1,
+          "min"    : 1
+        ],
+        "deploy.server.groups": [
+          "us-west-2": [
+            "app-test-v002"
+          ]
+        ]
+      )
     }
   }
 
@@ -53,7 +116,7 @@ class StageExpressionFunctionProviderSpec extends Specification {
     )
 
     when:
-    def currentStage = currentStage(pipeline)
+    def currentStage = StageExpressionFunctionProvider.currentStage(pipeline)
 
     then:
     currentStage.name == expectedStageName
@@ -72,7 +135,7 @@ class StageExpressionFunctionProviderSpec extends Specification {
     )
 
     when:
-    currentStage(pipeline)
+    StageExpressionFunctionProvider.currentStage(pipeline)
 
     then:
     thrown(SpelHelperFunctionException)
@@ -88,7 +151,7 @@ class StageExpressionFunctionProviderSpec extends Specification {
 
   def "stageByRefId() should match on #matchedAttribute"() {
     expect:
-    stageByRefId(pipeline, stageCriteria).name == expectedStageName
+    StageExpressionFunctionProvider.stageByRefId(pipeline, stageCriteria).name == expectedStageName
 
     where:
     stageCriteria || matchedAttribute || expectedStageName
@@ -98,9 +161,46 @@ class StageExpressionFunctionProviderSpec extends Specification {
 
   def "stageByRefId() should raise exception if stage not found"() {
     when:
-    stageByRefId(pipeline, "does_not_exist")
+    StageExpressionFunctionProvider.stageByRefId(pipeline, "does_not_exist")
 
     then:
     thrown(SpelHelperFunctionException)
+  }
+
+  @Unroll
+  def "stage() should match on #matchedAttribute"() {
+    expect:
+    StageExpressionFunctionProvider.stage(pipeline, stageCriteria).context.region == expectedRegion
+
+    where:
+    stageCriteria     || matchedAttribute || expectedRegion
+    "My Second Stage" || "name"           || "us-west-1"
+    "1"               || "id"             || "us-east-1"
+    "2"               || "id"             || "us-west-1"
+  }
+
+  def "stage() should raise exception if stage not found"() {
+    when:
+    StageExpressionFunctionProvider.stage(pipeline, "does_not_exist")
+
+    then:
+    thrown(SpelHelperFunctionException)
+  }
+
+  def "stageExists() returns whether a stage exists or not"() {
+    when:
+    def exists = StageExpressionFunctionProvider.stageExists(pipeline, id)
+
+    then:
+    exists == real
+
+    where:
+    id                   | real
+    "My First Stage"     | true
+    "1"                  | true
+    "My Second Stage"    | true
+    "2"                  | true
+    "Non-existent Stage" | false
+    "42"                 | false
   }
 }
