+++ orca-front50/src/main/groovy/com/netflix/spinnaker/orca/front50/pipeline/PipelineExpressionFunctionProvider.java
@@ -20,17 +20,14 @@ import static java.lang.String.format;
 
 import com.google.common.base.Strings;
 import com.netflix.spinnaker.kork.core.RetrySupport;
+import com.netflix.spinnaker.kork.expressions.ExpressionFunctionProvider;
+import com.netflix.spinnaker.kork.expressions.SpelHelperFunctionException;
 import com.netflix.spinnaker.orca.front50.Front50Service;
-import com.netflix.spinnaker.orca.pipeline.expressions.ExpressionFunctionProvider;
-import com.netflix.spinnaker.orca.pipeline.expressions.SpelHelperFunctionException;
 import com.netflix.spinnaker.orca.pipeline.model.Execution;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -38,11 +35,9 @@ public class PipelineExpressionFunctionProvider implements ExpressionFunctionPro
   // Static because it's needed during expression eval (which is a static)
   private static Front50Service front50Service = null;
 
-  @Autowired
   PipelineExpressionFunctionProvider(Optional<Front50Service> front50Service) {
-    if (front50Service.isPresent()) {
-      PipelineExpressionFunctionProvider.front50Service = front50Service.get();
-    }
+    front50Service.ifPresent(
+        service -> PipelineExpressionFunctionProvider.front50Service = service);
   }
 
   @Nullable
@@ -53,17 +48,16 @@ public class PipelineExpressionFunctionProvider implements ExpressionFunctionPro
 
   @NotNull
   @Override
-  public Collection<FunctionDefinition> getFunctions() {
-    return Arrays.asList(
+  public Functions getFunctions() {
+    return new Functions(
         new FunctionDefinition(
             "pipelineId",
-            Arrays.asList(
-                new FunctionParameter(
-                    Execution.class,
-                    "execution",
-                    "The execution containing the currently executing stage"),
-                new FunctionParameter(
-                    String.class, "pipelineName", "A valid stage reference identifier"))));
+            new FunctionParameter(
+                Execution.class,
+                "execution",
+                "The execution containing the currently executing stage"),
+            new FunctionParameter(
+                String.class, "pipelineName", "A valid stage reference identifier")));
   }
 
   /**
